requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
transformers==4.35.0
torch==2.1.0
numpy==1.24.3
scikit-learn==1.3.2
spacy==3.7.2
scispacy==0.5.1
requests==2.31.0
python-multipart==0.0.6
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
accelerate==0.24.0
